Update extract_gfc.R
Here I have parallelized this function a couple of ways. To start the parallelization process, the user has to specify "cluster" for the number of nodes in the cluster they are working with. If they specify nothing or 1 for "cluster", the function will proceed as normal. If 2 or above are specified,  then projectRaster function is parallelized if they wish to convert to UTM. Also,  If 2 or above are specified, the for loop is parallelized using the foreach function.  I have deliberately chosen socket parallelization which should work on windows and linux/unix. The user also has to install the "foreach" package for this to work.